+++ angular-workspace/projects/ni/nimble-angular/src/directives/anchor-tab/nimble-anchor-tab-router-link-with-href.directive.ts
@@ -1,5 +1,6 @@
 import { Directive, Input } from '@angular/core';
-import { RouterLinkWithHref } from '@angular/router';
+import type { AnchorTab } from './nimble-anchor-tab.directive';
+import { DisableableRouterLinkWithHrefDirective } from '../anchor-base/disableable-router-link-with-href.directive';
 
 /**
  * Selectors used for built-in Angular RouterLink directives:
@@ -15,7 +16,7 @@ import { RouterLinkWithHref } from '@angular/router';
  * won't also be an active RouterLink directive incorrectly handling navigation.
  */
 @Directive({ selector: 'nimble-anchor-tab[nimbleRouterLink]' })
-export class NimbleAnchorTabRouterLinkWithHrefDirective extends RouterLinkWithHref {
+export class NimbleAnchorTabRouterLinkWithHrefDirective extends DisableableRouterLinkWithHrefDirective<AnchorTab> {
     @Input()
     public set nimbleRouterLink(commands: never[] | string | null | undefined) {
         this.routerLink = commands;