+++ angular-workspace/projects/ni/nimble-angular/src/directives/toggle-button/nimble-toggle-button-control-value-accessor.directive.ts
@@ -1,11 +1,11 @@
 import { Directive, forwardRef } from '@angular/core';
-import { NG_VALUE_ACCESSOR, CheckboxControlValueAccessor } from '@angular/forms';
+import { NG_VALUE_ACCESSOR } from '@angular/forms';
+import { CheckboxControlValueAccessor } from '../../thirdparty/directives/checkbox_value_accessor';
 
 /**
  * Extension of Angular's CheckboxControlValueAccessor to target the Nimble toggle button control.
  *
- * Directive decorator based on CheckboxControlValueAccessor decorator
- * https://github.com/angular/angular/blob/bbababe5900ea8f4c8fccd88238f6fe08a2ceb63/packages/forms/src/directives/checkbox_value_accessor.ts#L42
+ * Directive decorator based on CheckboxControlValueAccessor decorator in thirdparty/directives/checkbox_value_accessor
  */
 @Directive({
     selector: