+++ angular-workspace/projects/ni/nimble-angular/src/thirdparty/testing/matchers.d.ts
@@ -1,0 +1,6 @@
+// Manually created for Nimble. Matchers interface should include any functions defined in the object being passed to jasmine.addMatchers() in ./matchers.ts.
+declare namespace jasmine {
+    interface Matchers<T> {
+        toHaveText(expected: string): boolean;
+    }
+}